requirements.txt
requests
pillow
openai
python-dotenv
python-dotenv
google-cloud-vision